% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data set}
\format{
Data frame containing 2 variables and 15 rows.
}
\source{
Canvas
}
\usage{
fire
}
\description{
A data set provided to us for use in a lab assignment.
}
\keyword{datasets}
