% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Sample size}

\item{p}{Probability of success}
}
\value{
Creates plot for binomial simulation
}
\description{
mybin
}
\examples{
\dontrun{mybin(iter = 100, n = 10, p = 0.7)}
}
