% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{lambda}{Value of lambda (iter/n)}

\item{...}{Additional arguments, if needed}
}
\value{
A very pretty set of graphs for the distribution
}
\description{
mycltp
}
\examples{
\dontrun{mycltp(n = 2, iter = 10000, lambda = 4)}
}
