% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt}
\usage{
myddt(df, SPECIES, col)
}
\arguments{
\item{df}{The data file to pull from (must be a .csv file)}

\item{SPECIES}{The species you wish to filter by}

\item{col}{The color you wish to make the plot's points (variable)}
}
\value{
Produces a plot of length vs weight for declared species, table of relative frequency values per river, and tables of species-filtered and unfiltered data
}
\description{
myddt
}
\examples{
\dontrun{myddt(df = ddt, SPECIES = "CCATFISH")}
}
