% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{a}{Upper tail value}
}
\value{
Finds area under normal curve
}
\description{
myncurve
}
\examples{
\dontrun{myncurve(mu = 3, sigma = 2, a = 4)}
}
