% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R, R/mymaxlik.R
\docType{data}
\name{x}
\alias{x}
\title{Title}
\format{
An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.
}
\usage{
x

x
}
\arguments{
\item{iter}{Number of iterations}

\item{fun}{Function to evaluate with}

\item{alpha}{Bootstrap interval value (95% = 0.05)}

\item{lfun}{The function to use (binomial, poisson, etc)}

\item{x}{Vector of values}

\item{param}{Parameter to calculate (prob, theta, lambda, etc)}

\item{...}{Additional arguments, if needed}
}
\value{
Provides a histogram of sample evaluation using requested function

Provides index values for different variables, and a graph illustrating the maximum likelihood value
}
\description{
Title

mymaxlik
}
\examples{
\dontrun{myboot2(iter=10000,sam,fun="mean",alpha=0.30,cx=1.5)}
\dontrun{mymaxlik(x = c(3, 3, 4, 3, 4, 5, 5, 4), param = seq(0, 1, length = 1000), lfun = function(x, param) log(dbinom(x,prob=param,size=10)), xlab = expression(pi), main = "Binomial, n = 20")}
}
\keyword{datasets}
